requirements.txt
# Core Dependencies
python>=3.11

# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0

# Groq Integration
groq>=0.4.0
langchain>=0.1.0
langchain-groq>=0.0.1
instructor>=0.4.0

# Memory & Storage
redis>=5.0.0
chromadb>=0.4.18
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.23

# Code Operations
GitPython>=3.1.40
tree-sitter>=0.20.4
black>=23.12.0
ruff>=0.1.8
mypy>=1.7.1

# Utilities
chardet>=5.2.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
aiofiles>=23.2.1

# Monitoring & Observability
prometheus-client>=0.19.0
sentry-sdk>=1.39.1
langfuse>=2.0.0

# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0
httpx>=0.25.2

# Development
ipython>=8.18.1
jupyter>=1.0.0
